stuff i did in class today. not complete, need to finish the _update_animation method
